+++ programs/artiste/src/artiste_desktop.h
@@ -136,7 +136,7 @@ public:
     font.unload();
   }
 
-  bool input_key_down(signed long key)
+  bool input_key_down(uint16_t key)
   {
     if (!k_focus)
       return true;
@@ -226,7 +226,7 @@ public:
     return false;
   }
 
-  bool input_key_up(signed long key)
+  bool input_key_up(uint16_t key)
   {
     if (!k_focus)
       return true;
